Don't retain artifacts longer than a day (#78)
Add `retention-days: 1` to all builds.

We upload artifacts so we can publish them to itch and gh-pages, but we don't really need the builds after that. You can quickly run into the 0.75 GB limit from Github free for artifacts that you don't need.

While it can also be useful to download artifacts from the Actions page, it seems better to err on the side of making it easy to increase lifetime instead of no clear lifetime.